Added user authorization to only allows users who created a post to edit/delete it

+++ backend/routes/posts.js
@@ -42,7 +42,8 @@ router.post(
 		const post = new Post({
 			title: req.body.title,
 			content: req.body.content,
-			imagePath: url + '/images/' + req.file.filename
+			imagePath: url + '/images/' + req.file.filename,
+			creator: req.userData.userId
 		});
 		post.save().then(createdPost => {
 			res.status(201).json({
@@ -56,6 +57,7 @@ router.post(
 		})
 	}
 );
+
 
 router.put(
 	'/:id',
@@ -71,15 +73,18 @@ router.put(
 			_id: req.body.id,
 			title: req.body.title,
 			content: req.body.content,
-			imagePath: imagePath
+			imagePath: imagePath,
+			creator: req.userData.userId
 		});
-		Post.updateOne({ _id: req.params.id }, post).then(result => {
-			console.log(result);
-			res.status(200).json({ message: "Update Sucessful!" });
+		Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+			if (result.nModified > 0) {
+				res.status(200).json({ message: "Update Sucessful!" });
+			} else {
+				res.status(401).json({ message: "Update NOT Successful. Not Authorized" });
+			}
 		});
 	}
 );
-
 
 
 router.get(
@@ -117,7 +122,6 @@ router.get(
 );
 
 
-
 router.get('/:id', (req, res, next) => {
 	Post.findById(req.params.id).then((post) => {
 		if (post) {
@@ -128,10 +132,16 @@ router.get('/:id', (req, res, next) => {
 	});
 });
 
+
 router.delete("/:id", checkAuth, (req, res, next) => {
-	Post.deleteOne({ _id: req.params.id }).then(result => {
+	Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then( result => {
+		// result.nModified not present in the json with deletion
 		console.log(result);
-		res.status(200).json({ message: "Post deleted!" });
+		if (result.n > 0) {
+			res.status(200).json({ message: "Deletion Successfull" });
+		} else {
+			res.status(401).json({ message: "Deletion NOT Successful. Not Authorized" });
+		}
 	});
 });
